Fix/90

+++ src/main/java/com/example/backend/controller/PinController.java
@@ -52,6 +52,9 @@ public class PinController {
 
     @DeleteMapping("/pin/deletePin/{pin_id}")
     public CommonResult deletePinById(@PathVariable(value = "pin_id", required = false) Integer pin_id) {
+        int photoRet = photoService.deletePhotoByPinId(pin_id);
+        if (photoRet == 0)
+            return CommonResult.failed("删除外键p_id = " + pin_id + "的photo失败");
         int ret = pinService.deletePinById(pin_id);
         if (ret == 0)
             return CommonResult.failed("删除id = " + pin_id + "的pin失败");
